+++ bdranalytics/pdlearn/tests/test_pipeline.py
@@ -1,9 +1,10 @@
-from bdranalytics.pandaspipeline.transformers import PdLagTransformer, PdWindowTransformer, PdFeatureUnion, \
-    PdFeatureChain
-import unittest
-import pandas as pd
 import numpy as np
+import pandas as pd
+import unittest
 from sklearn.pipeline import FeatureUnion, Pipeline
+
+from bdranalytics.pdlearn.pipeline import PdFeatureUnion, PdFeatureChain
+from bdranalytics.pdlearn.preprocessing import PdLagTransformer, PdWindowTransformer, PdFeatureUnion
 
 
 class TestLagTransformer(unittest.TestCase):
