+++ bdranalytics/pdlearn/tests/test_preprocessing.py
@@ -1,10 +1,9 @@
-import unittest
-
 import numpy as np
 import pandas as pd
+import unittest
 
-from bdranalytics.pdpreprocessing.encoders import DateCyclicalEncoding, DateOneHotEncoding
-from bdranalytics.pdpreprocessing.encoders import date_to_dateparts, date_to_cyclical
+from bdranalytics.pdlearn.preprocessing import DateCyclicalEncoding, DateOneHotEncoding
+from bdranalytics.pdlearn.preprocessing import date_to_dateparts, date_to_cyclical
 
 
 class TestDatePartitioner(unittest.TestCase):
